layout/style/test/chrome/test_chrome_only_media_queries.html

<!doctype html>
<title>Test for parsing of non-content-exposed media-queries.</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<style></style>
<script>
const SHEET = document.querySelector('style');

function expect(q, shouldBeParseable) {
  const NOT_PARSEABLE_TEXT = "@media screen, not all {\n}";

  let text = "@media screen, " + q + " {\n}";
  SHEET.textContent = text;

  let rules = SHEET.sheet.cssRules;
  is(rules.length, 1, `Rule not parsed for ${q}`);
  is(rules[0].cssText,
     shouldBeParseable ? text : NOT_PARSEABLE_TEXT,
     `Serialization for ${q}`);
}

function expectParseable(q) {
  expect(q, true);
}

function expectNonParseable(q) {
  expect(q, false);
}

// Test a toggle that should always match for `1` or `0`.
function testToggle(toggle) {
  expectParseable(`(${toggle})`);
  expectParseable(`(${toggle}: 1)`);
  expectParseable(`(${toggle}: 0)`);

  expectNonParseable(`(${toggle}: foo)`);
  expectNonParseable(`(min-${toggle}: 0)`);
  expectNonParseable(`(max-${toggle}: 0)`);
  expectNonParseable(`(max-${toggle})`);
  expectNonParseable(`(min-${toggle})`);

  let matches_1 = matchMedia(`(${toggle}: 1)`).matches;
  let matches_0 = matchMedia(`(${toggle}: 0)`).matches;
  isnot(matches_0, matches_1, `Should not match both true and false: ${toggle}`);
  is(matches_0 || matches_1, true, `Should match at least one: ${toggle}`);
}

const TOGGLES = [
  "-moz-is-glyph",
  "-moz-print-preview",
  "-moz-scrollbar-start-backward",
  "-moz-scrollbar-start-forward",
  "-moz-scrollbar-end-backward",
  "-moz-scrollbar-end-forward",
  "-moz-overlay-scrollbars",
  "-moz-windows-default-theme",
  "-moz-mac-graphite-theme",
  "-moz-mac-yosemite-theme",
  "-moz-windows-accent-color-in-titlebar",
  "-moz-windows-compositor",
  "-moz-windows-classic",
  "-moz-windows-glass",
  "-moz-swipe-animation-enabled",
];

for (let i = 0; i < TOGGLES.length; ++i) {
  testToggle(TOGGLES[i])
}

expectParseable("(-moz-os-version: windows-win7)");
expectParseable("(-moz-os-version: windows-win8)");
expectParseable("(-moz-os-version: windows-win10)");
expectNonParseable("(-moz-os-version: )");

expectParseable("(prefers-contrast: more)")
expectParseable("(prefers-contrast: forced)")
expectParseable("(prefers-contrast: no-preference)")
expectParseable("(prefers-contrast: less)");
expectParseable("(prefers-contrast)")
</script>
